refactor VideoCallScreen
